+++ extensions/hotflow/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Hotflow Changelog
 
-## [Initial Version] - 2026-01-15
+## [Initial Version] - {PR_MERGE_DATE}
 
 - Initial Version
-- Support GitHub, Hacker News, Product Hunt, X as sources for trending content.
+- Support GitHub, Hacker News, Product Hunt, X, BiliBili as sources for trending content.